Checks for User Interrupt

Check if the user has requested an interrupt, and if so branch to R's error signaling functions.

R_CheckUserInterrupt(void)

Source

Declaration: src/include/R_ext/Utils.h

Implementation: src/main/errors.c

Arguments

void

(void) Nothing.

Value

Nothing.

C API

#include <R_ext/Utils.h>

void R_CheckUserInterrupt(void)

R API

All evaluation at the top level of R checks for interrupts.

Related functions that affect R_CheckUserInterrupt():

base::suspendInterrupts() - suspends interrupts during evaluation

base::allowInterrupts() - allows interrupts during evaluation